Native quaternion implementation
This PR removes the dependency on boost::math::quaternion in favor of a minimal
native implementation, which includes simple arithmetic and comparison
operators, and conj, abs, norm and pow function implementations.

The class also provides versor() and is_versor() methods for creating and
testing for unit quaternions, used in many pointing operations in the maps
package.

The G3Quat object is a true serializeable G3FrameObject with a numpy buffer
interface.  The derived G3VectorQuat and G3MapQuat classes maintain backward
compatibility for deserializing data stored on disk.