Consolidate IO stream handling in one place (#164)
This PR migrates all of the code that handles file and buffer IO stream creation
into one place.  This ensures consistent handling of all input and output data
serialization streams, and minimizes the footprint of the Boost iostreams and
filesystem library in the core code, making it incrementally easier to
eventually replace Boost altogether.