Avoid bounds checking in libc++ std::vector when setting ranges.
basic_streambuf<char> needs char points to define the range of the buffer,
which we get by getting pointers to the start and end. end pointers, like
end buffers, are invalid entries: they are after the last element. Newer
libc++ asserts if you asks for elements after the end of a vector, however.
Just ask for the 0th element and add size, which does the same thing.