Consistent G3Frame versioning and serialization (#181)
This PR uses the same versioning and archive serialization methods for G3Frame instances as those for all other serializable objects in the library.

The G3Frame class version is defined in the header, along with specialized load/save methods for IO with cereal binary archives.

The load/save methods for streams are renamed to loads/saves to ensure that the cereal library binds the correct serialization methods.

This PR does not change the serialized blob structure or increment the serialization version, so full compatibility with existing data is maintained.