Shim class for loading G3SuperTimestream objects from disk
This PR is the first step in integrating the G3SuperTimestream class used by the
Simons Observatory team into the core library.  This class is a very thin
wrapper around the core G3TimestreamMap class.  It is designed only to load
G3SuperTimestream objects from G3 files on disk for immediate conversion to
G3TimestreamMap objects.  It is also not exposed to the public or python API.