Preserve insertion order in G3TimestreamMap
Use a custom OrderedMap class that behaves identically to std::map, except that
items are iterated over in the order they were inserted into the map.

This allows preserving the detector ordering in converting between
G3SuperTimestream and G3TimestreamMap objects, as the former are not guaranteed
to be ordered alphabetically.