Preserve insertion order in G3TimestreamMap
Use a custom OrderedMap class that behaves identically to std::map, except that
items are iterated over in the order they were inserted into the map.

This allows preserving the detector ordering when creating a G3TimestreamMap
from a 2D array and a set of keys, e.g. when converting between
G3SuperTimestream and G3TimestreamMap objects, as the former are not guaranteed
to be ordered alphabetically.