Add support for full-depth compression and decompression of int32 timestreams
Add a bit_depth attribute to G3Timestream objects, allowing the flac compression algorithm
to enable compression of the full 32 bits per sample, rather than restricting to 24 bits.

If 32-bit compression is selected, then int64 timestreams are truncated to int32, and
double timestreams are truncated to float prior to compression, and the resulting
data type is stored along with the compressed data.