Memory-efficient variance for G3TimestreamMap objects
The numpy ndarray var() implementation involves an array copy operation that
doubles memory usage, which can drive the memory requirements for timestream
processing.  This PR adds a var() implementation that avoids this copy operation
for the relatively common case of efficiently computing the variance of each
timestream in a G3TimestreamMap.  This code is called automatically when using
the `numpy.var()` function over the last axis on a timestream map object.