Fix function resolution in docs in python 3.13+
The .__module__ attribute for class methods is used differently in Python 3.13
than it was in older python versions, perhaps due to divergence in how
Boost.python handles the object model in constructing classes.  Work around this
by avoiding recursion through class objects (which have a .mro() method) when
searching for functions to document.