Fix gzip compression
The GZip compression implementation used the base ZLib stream functions and did
not account for the necessary GZip header and footer.  This PR reimplements the
GZipDecoder and GZipEncoder using the correct gzread/gzwrite and associated
functions.

This code has been verified to be able to read gzip-compressed files written
by the original Boost.iostreams implementation.