Migration to Pybind11 (#197)
This PR migrates all of the library code to `pybind11`, while striving to maintain the same feature set that was provided by Boost.Python.

In the meantime, the master branch has been updated with some of the more general features in this PR, and to include a shim structure that makes much of the boost binding code (e.g. method argument handling) look more like the pybind11 API; see `core/pybind11_compat.h` for details.  This was done to minimize the diff in this PR to focus on the important changes.

The changes to focus on are summarized here.

### Build system overhaul

Pybind11 is downloaded automatically by cmake using the `FetchContent` framework, or installed by pip.  We build against the latest version (2.13.6) since this is an easy dependency to get.  Linking to both pybind11 and python is done using the link targets provided with the pybind11 library (`pybind11::module` and `pybind11::embed`), which removes the vast majority of hand-holding we've had to do with Boost.

The documentation has been updated to reflect the new framework.

### Class bindings

Objects are bound with `std::shared_ptr` holder types to allow sharing memory between C++ and Python with persistent scope, and enable dynamic attribute assignment for duck-typing from Python.

Pointer conversions between derived classes and their bases are handled automatically and do not require specific registration.  On the other hand, conversions from python objects to some types (enums, containers) are handled with appropriate constructors and the `implicitly_convertible` function.

Type resolution seems to work more accurately with pybind11, so many of the "catch-all" member functions that required casting python objects to various types have been scrapped in favor of several smaller targeted overloads, letting the compiler deal with the logic.

Bound classes are non-copyable by default, unless a copy constructor is added to the bindings.  To ensure that loggers and pipeline modules are not copy-constructible also in C++, the copy constructor and assignment operators are deleted in the base classes.

The pybind11 pickling framework requires access to the default constructor for serializable objects, similarly to cereal.  To enable this, the `G3Pickler` class can be added as a friend for classes where the default constructor is private.

Many of the little set / get functions that were defined statically with boost have been replaced with C++ lambda functions.

### Frames, Modules and Pipelines

The pipeline infrastructure remains nearly identical with pybind11.  Registration functions are provided for classes derived from G3FrameObject and G3Module.  Interoperability between python and C++ is maintained with a G3PythonModule (formerly G3ModuleWrap) trampoline class that works in much the same way as the boost wrapper class.

### Visibility

The pybind11 namespace is marked with hidden visibility.  To enable the shared library linking mechanism to work, the `G3ModuleRegistrator` class is marked with default visibility.

A handful of C++ pipeline modules that include python structures (typically callback functions) as members are also marked with hidden visibility.  These classes are not part of the public API, so this should be fine, assuming someone isn't trying to uses these outside of a python environment.  If necessary, this can be avoided by using trampoline classes in the same way as is done for the G3PythonModule class.

### Vectors and Maps

Container types are implemented opaquely using the machinery in `pybind11/stl_bind.h` rather than relying on the default `pybind11/stl.h` implementation.  The `bind_vector` and `bind_map` functions are deconstructed to use shared pointers by default, and to enable overriding the default buffer protocol implementation.

Crucially, these bindings are *not* made module-local to enable importing them in each module without re-registration.  Technically this means that this package cannot be loaded together with other packages that also define global bindings for these STL types, but the likelihood of running into such a package seems relatively low.  Doing otherwise requires some gymnastics with wrapper classes and type casters, and/or exposing these types directly as numpy arrays.

Implementation details are all in the `core/container_pybindings.h` header.

Bound vectors have from-array constructors that enable type conversion by copy, and use the greatly-simplified `buffer_info` structure to enable views -- see for example the G3Vector, G3Timestream and G3Quat implementations.  A simple custom `cbuffer` class is used to ensure that python objects are contiguous before they even reach the bindings, which greatly simplifies the amount of handling required in the code.  Additionally, the numpy interface enables straightforward type conversion and yet simpler buffer handling, without having to link against the numpy library headers.

Maps return iterator-view-like objects for `.keys()`, `.values()` and `.items()` methods, rather than python lists.  The special reference-by-proxy item handling that boost required is (it seems) handled under the hood by pybind11.  G3Frame objects also return iterator-view-like objects for `.keys()`, `.values()` and `.items()`, and include an iterator over the key names on the C++ side.

---------

Co-authored-by: Nathan Whitehorn <nwhitehorn@pa.msu.edu>