Fix construction of G3FrameObject maps and vectors
Pybind11 performs a series of type conversions when passing python arguments to C++
functions.  These conversions do not behave appropriately when the argument is a base
G3FrameObject.  This commit overrides all of the functions in the G3MapFrameObject
and G3VectorFrameObject classes that take G3FrameObject objects as input arguments.
Type conversion is disabled on any argument that is a G3FrameObject, and the
overridden functions are used in place of the standard ones.