Bug fix for python bindings for non-contiguous input vectors
When a non-contiguous numpy vector (e.g. a sliced view into a numpy array) is
used as an input argument to a bound function, the underlying C++ object does
not properly account for strides in the input array.  Use the pybind11 numpy
interface to handle this properly, while still allowing type conversion.